refactor(#582): inject the tick-loop clock and home the test stand-ins
The worker runtime takes a now option beside timestep and client, so the
fast-clock test stand-in drives the tick loop by injection instead of
patching the global performance.now — the restore-ordering footgun goes
with it. The fail-first transport handler moves to test-utils per the
stand-in convention, and the authed-client helper dissolves into the
per-test inline sequence its sibling suites already use.
